fix(hydration): escape css var name to avoid mismatch
